Refactored with better components.
