+++ src/styles/stylesSearchBar.js
@@ -1,17 +1,5 @@
 import styled from "styled-components";
 
-export const MovieContainer = styled.div`
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-
-    .no-movie,
-    .loading {
-        padding: 1rem;
-        font-size: 3rem;
-    }
-`;
-
 export const SearchHeader = styled.header`
     display: flex;
     justify-content: end;
